Bugfix: Dynamic form fixes to small rounding errors in form height for forms with very large numbers of fields.

+++ release2/src/org/yawlfoundation/yawl/resourcing/jsf/dynform/SubPanelController.java
@@ -215,7 +215,7 @@ public class SubPanelController {
     public SubPanel addSubPanel(SubPanel newPanel) {
 
         // adjust the tops of subsequent panels at the same depth level
-        int adjustment = newPanel.getHeight() + DynFormFactory.Y_DEF_INCREMENT;
+        int adjustment = (int) newPanel.getHeight() + DynFormFactory.Y_DEF_INCREMENT;
         int top = newPanel.getTop();
         incSubPanelTops(top, adjustment) ;
 
@@ -246,7 +246,7 @@ public class SubPanelController {
      * @return the outermost containing subpanel of the one removed
      */
     public SubPanel removeSubPanel(SubPanel oldPanel) {
-        int adjustment = - (oldPanel.getHeight() + DynFormFactory.Y_DEF_INCREMENT);
+        int adjustment = - ((int) oldPanel.getHeight() + DynFormFactory.Y_DEF_INCREMENT);
         int top = oldPanel.getTop();
 
         incSubPanelTops(top, adjustment) ;
